refactor out streamlit server start
