+++ lib/pymedphys/_gui.py
@@ -15,13 +15,8 @@
 
 import pathlib
 import shutil
-import uuid
-from typing import Any, Dict, Tuple
 
-from pymedphys._imports import streamlit as st
-from pymedphys._imports import tornado
-
-from pymedphys._streamlit.server import downloads
+from pymedphys._streamlit.server import start
 
 HERE = pathlib.Path(__file__).parent.resolve()
 STREAMLIT_CONTENT_DIR = HERE.joinpath("_streamlit")
@@ -34,62 +29,11 @@ def main(args):
     streamlit_script_path = str(HERE.joinpath("_app.py"))
 
     if args.port:
-        st.cli._apply_config_options_from_cli({"server.port": args.port})
+        config = {"server.port": args.port}
+    else:
+        config = {}
 
-    # Needs to run after config has been set
-    _monkey_patch_streamlit_server()
-
-    st._is_running_with_streamlit = True
-    st.bootstrap.run(streamlit_script_path, "", [])
-
-
-Handlers = Dict[str, Tuple[Any, Dict[str, Any]]]
-
-
-def _create_handlers() -> Handlers:
-    class HelloWorldHandler(  # pylint: disable = abstract-method
-        tornado.web.RequestHandler
-    ):
-        def get(self):
-            self.write("Hello world!")
-
-    class DownloadHandler(  # pylint: disable = abstract-method
-        tornado.web.RequestHandler
-    ):
-        def get(self, session_id: uuid.UUID, name: str):
-            file_bytes = downloads.get_download_file(session_id, name)
-
-            self.write(file_bytes)
-            self.finish()
-
-    return {
-        "pymedphys": (HelloWorldHandler, {}),
-        "downloads/(.*)/(.*)": (DownloadHandler, {}),
-    }
-
-
-def _monkey_patch_streamlit_server():
-    """Adds custom URL routes to Streamlit's tornado server."""
-    handlers = _create_handlers()
-
-    OfficialServer = st.server.server.Server
-    official_create_app = OfficialServer._create_app
-
-    def patched_create_app(self: st.server.server.Server) -> tornado.web.Application:
-        app: tornado.web.Application = official_create_app(self)
-
-        base: str = st.config.get_option("server.baseUrlPath")
-
-        rules: tornado.routing._RuleList = []
-        for key, (handler, kwargs) in handlers.items():
-            pattern = st.server.server_util.make_url_path_regex(base, key)
-            rules.append((pattern, handler, kwargs))
-
-        app.add_handlers(".*", rules)
-
-        return app
-
-    OfficialServer._create_app = patched_create_app
+    start.start_streamlit_server(streamlit_script_path, config)
 
 
 def _fill_streamlit_credentials():
